requirements.txt
streamlit
google-genai
chromadb
pypdf
pandas